% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_carousel_item.R
\name{owl_carousel_item}
\alias{owl_carousel_item}
\title{Title}
\usage{
owl_carousel_item(
  id,
  title,
  subtitle,
  content = NULL,
  button_text = NULL,
  width = "300px"
)
}
\arguments{
\item{title}{title}

\item{subtitle}{subtitle}

\item{content}{content}

\item{button_text}{button}

\item{width}{Width, in the format "300px" (default value)}
}
\description{
Title
}
