% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyslider.R
\name{tinyslider-shiny}
\alias{tinyslider-shiny}
\alias{tinysliderOutput}
\alias{renderTinyslider}
\title{Shiny bindings for tinyslider}
\usage{
tinysliderOutput(outputId, width = "100\%", height = "400px")

renderTinyslider(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a tinyslider}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using tinyslider within Shiny
applications and interactive Markdown documents.
}
\examples{
\dontrun{
library(shiny)
library(tinyslider)

ui <- fluidPage(
  tinysliderOutput("test")
)
server <- function(input, output, session) {

  output$test <- renderTinyslider({
    tinyslider(
      tinyslider_card(
        title = "First title",
        subtitle = "First subtitle",
        content = paste0("Lorem Ipsum is simply dummy text of the",
        "printing and typesetting industry. Lorem Ipsum has been the",
         "industry's standard dummy text ever since the 1500s, when",
         "an unknown printer took a galley of type and scrambled",
          "it to make a type specimen book."),
        image = "https://placeimg.com/200/150/nature"
      ),
      tinyslider_card(
        "Second title",
        "Second subtitle",
        "There is a button below",
        button_text = "See more",
        image = "https://placeimg.com/200/150/nature/2"
      ),
      tinyslider_card(
        "Third title",
        subtitle = NULL,
        content = paste0("Lorem Ipsum is simply dummy text of the",
        "printing and typesetting industry. Lorem Ipsum has been",
        "the industry's standard dummy text ever since the",
        "1500s, when an unknown printer took a galley of type and",
        "scrambled it to make a type specimen book. It has",
        "survived not only five centuries, but also the leap into",
        "electronic typesetting, remaining essentially unchanged.",
        "It was popularised in the 1960s with the",
        "release of Letraset sheets containing Lorem Ipsum passages,",
        "and more recently with desktop publishing",
        "software like Aldus PageMaker including versions of Lorem Ipsum.")
      ),
      options = list(
        loop = TRUE,
        autoplay = TRUE,
        autoplayTimeout = 1000
      )
    )
  })
}

shinyApp(ui, server, options = list(launch.browser = TRUE))
}

}
