% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyslider_card.R
\name{tinyslider_cards}
\alias{tinyslider_cards}
\alias{tinyslider_card}
\alias{tinyslider_card_2}
\title{Create cards for the tinyslider carousel}
\usage{
tinyslider_card(
  title = NULL,
  subtitle = NULL,
  content = NULL,
  button_text = NULL,
  image = NULL,
  width = NULL
)

tinyslider_card_2(
  title = NULL,
  subtitle = NULL,
  content = NULL,
  button_text = NULL,
  image = NULL,
  width = NULL
)
}
\arguments{
\item{title}{Title of the card}

\item{subtitle}{Subtitle of the card}

\item{content}{Text of the card}

\item{button_text}{Display some text inside a button placed at the bottom of the card. If \code{NULL}, the button is not created.}

\item{image}{An image to place on top of the card.}

\item{width}{Width of the card as a character, in px or in \%.}
}
\value{
A card in HTML code.
}
\description{
Create cards for the tinyslider carousel
}
\details{
\code{tinyslider_card} and \code{tinyslider_card_2} are very similar. The only difference is in the layout of the card. The latter creates a card with two columns. If \code{image} is not \code{NULL}, \code{tinyslider_card} puts an image at the top of the card, and \code{tinyslider_card_2} puts an image in the left column.
}
\examples{
\dontrun{
tinyslider_card("This is a title", "This is a subtitle")}
}
