% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyslider_item.R
\name{tinyslider_item}
\alias{tinyslider_item}
\title{Create items for the tinyslider carousel}
\usage{
tinyslider_item(
  id,
  title = NULL,
  subtitle = NULL,
  content = NULL,
  button_text = NULL,
  image = NULL,
  width = NULL
)
}
\arguments{
\item{id}{Id of the card}

\item{title}{Title of the card}

\item{subtitle}{Subtitle of the card}

\item{content}{Text of the card}

\item{button_text}{Display some text inside a button placed at the bottom of the card. If \code{NULL}, the button is not created.}

\item{image}{An image to place on top of the card.}
}
\description{
Create items for the tinyslider carousel
}
