% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_to_card}
\alias{plot_to_card}
\title{Convert a plot to a card}
\usage{
plot_to_card(gg)
}
\arguments{
\item{gg}{the name of a plot (works only for ggplots for now)}
}
\description{
Convert a plot to a card
}
