% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyslider.R
\name{tinyslider}
\alias{tinyslider}
\title{Create a carousel with tinyslider}
\usage{
tinyslider(id, ..., options = list(), width = NULL, height = NULL)
}
\arguments{
\item{id}{Id of the carousel}

\item{...}{Items created with \code{tinyslider_card}}

\item{options}{List of options for the carousel}
}
\description{
Create a carousel with tinyslider
}
