% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyslider.R
\name{tinyslider}
\alias{tinyslider}
\title{Create a carousel with tinyslider}
\usage{
tinyslider(
  elementId = NULL,
  ...,
  options = list(),
  responsive_options = list()
)
}
\arguments{
\item{elementId}{Id of the carousel}

\item{...}{Items created with \code{tinyslider_card()}, or plots embedded in \code{plot_to_card}.}

\item{options}{List of options for the carousel.}

\item{responsive_options}{List of options that differ according to viewport size. See the examples.}
}
\value{
An htmlwidget visible in RStudio viewer or in a web browser. This htmlwidget can be stored into an object (such as \code{x <- tinyslider()}.
}
\description{
Create a carousel with tinyslider
}
\details{
Due to its length, I do not detail the list of options available. This list is available \href{https://github.com/ganlanyuan/tiny-slider#options}{here}.

WARNING: the option "container" is automatically set in the function and therefore should not be modified.
}
\examples{
\dontrun{
tinyslider(
  tinyslider_card(
    title = "First title",
    subtitle = "First subtitle",
    content = paste0(
      "Lorem Ipsum is simply dummy text of",
      "the printing and typesetting industry. Lorem Ipsum has been the",
      "industry's standard dummy text ever since the 1500s,",
      "when an unknown printer took a galley of type and scrambled",
      "it to make a type specimen book."
    ),
    image = "https://placeimg.com/200/150/nature"
  ),
  tinyslider_card(
    "Second title",
    "Second subtitle",
    "There is a button below",
    button_text = "See more",
    image = "https://placeimg.com/200/150/nature/2"
  ),
  tinyslider_card(
    "Third title",
    subtitle = NULL,
    content = paste0(
      "Lorem Ipsum is simply dummy text of",
      "the printing and typesetting industry. Lorem Ipsum has been",
      "the industry's standard dummy text ever since the 1500s,",
      "when an unknown printer took a galley of type and",
      "scrambled it to make a type specimen book. It has",
      "survived not only five centuries, but also the leap into",
      "electronic typesetting, remaining essentially unchanged.",
      "It was popularised in the 1960s with the",
      "release of Letraset sheets containing Lorem Ipsum passages,",
      "and more recently with desktop publishing",
      "software like Aldus PageMaker including versions of Lorem Ipsum."
    )
  ),
  options = list(
    loop = TRUE,
    autoplay = TRUE,
    autoplayTimeout = 1000
  )
)
}

}
